+++ src/net/tonkovich/java/Discount20.java
@@ -4,7 +4,7 @@ public class Discount20 implements DiscountOptions {
   
   private double percentage = 0.20;
 
-  public double calculate(double price, double quantity) {
+  public double calculate(double price, int quantity) {
     double original = price*quantity;
     double subtract = original * percentage;
     return original - subtract;
