+++ src/net/tonkovich/java/DiscountOptions.java
@@ -2,7 +2,7 @@ package net.tonkovich.java;
 
 public interface DiscountOptions {
   
-  double calculate(double price, double quantity);
+  double calculate(double price, int quantity);
   
   double getPercentage();
 }
